# public/denis23x.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "version": "1.0.0",
    "description": "This API provides access to a collection of books, users, and reviews, allowing clients to perform various operations like viewing available books, creating and managing users, and posting or retrieving book reviews.\n\n- **Books**: Retrieve book details, search for books, and view books.\n- **Users**: Manage users, view user profiles.\n- **Reviews**: Add, view, update, or delete reviews for books.",
    "title": "Books Showcase Demo",
    "contact": {
      "email": "denix23x@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
...
  "paths": {
...
    "/api/v1/placeholder/users/{id}": {
      "get": {
        "summary": "Get User",
        "deprecated": false,
        "description": "",
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": 1,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/User"
                    },
                    "status": {
                      "type": "integer"
                    }
                  }
                },
                "examples": {
                  "1": {
                    "summary": "Success",
                    "value": {
                      "data": {
                        "id": 970,
                        "avatar": "https://picsum.photos/seed/RA0jOE/512/512?blur=1",
                        "firstName": "Dangelo",
                        "lastName": "Vandervort-Ryan",
                        "gender": "Non-binary",
                        "sex": "Male",
                        "bio": "veteran, educator",
                        "email": "King.Fadel@gmail.com",
                        "createdAt": "2024-04-30T01:45:25.033Z",
                        "updatedAt": "2024-11-14T12:32:36.313Z"
                      },
                      "status": 200
                    }
                  }
                }
              }
            },
            "headers": {}
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {}
                    },
                    "status": {
                      "type": "integer"
                    }
                  }
                },
                "examples": {
                  "1": {
                    "summary": "Exception",
                    "value": {
                      "error": {
                        "name": "PrismaClientKnownRequestError",
                        "code": "P2025",
                        "clientVersion": "5.22.0",
                        "meta": {
                          "modelName": "modelName",
                          "cause": "Record not found."
                        }
                      },
                      "status": 404
                    }
                  }
                }
              }
            },
            "headers": {}
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {}
                    },
                    "status": {
                      "type": "integer"
                    }
                  }
                },
                "examples": {
                  "1": {
                    "summary": "Exception",
                    "value": {
                      "error": {
                        "name": "PrismaError",
                        "clientVersion": "5.22.0"
                      },
                      "status": 500
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "delete": {
        "summary": "Delete User",
        "deprecated": false,
        "description": "",
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": 1,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/User"
                    },
                    "status": {
                      "type": "integer"
                    }
                  }
                },
                "examples": {
                  "1": {
                    "summary": "Success",
                    "value": {
                      "data": {
                        "id": 289,
                        "avatar": "https://picsum.photos/seed/FnIvGFmuX/512/512",
                        "firstName": "Geovanni",
                        "lastName": "Dooley",
                        "gender": "Two* person",
                        "sex": "Male",
                        "bio": "cornet junkie, model",
                        "email": "Charity_Bosco18@yahoo.com",
                        "createdAt": "2024-10-17T10:48:43.231Z",
                        "updatedAt": "2024-11-14T04:55:25.080Z"
                      },
                      "status": 200
                    }
                  }
                }
              }
            },
            "headers": {}
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {}
                    },
                    "status": {
                      "type": "integer"
                    }
                  }
                },
                "examples": {
                  "1": {
                    "summary": "Exception",
                    "value": {
                      "error": {
                        "name": "PrismaClientKnownRequestError",
                        "code": "P2025",
                        "clientVersion": "5.22.0",
                        "meta": {
                          "modelName": "modelName",
                          "cause": "Record not found."
                        }
                      },
                      "status": 404
                    }
                  }
                }
              }
            },
            "headers": {}
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {}
                    },
                    "status": {
                      "type": "integer"
                    }
                  }
                },
                "examples": {
                  "1": {
                    "summary": "Exception",
                    "value": {
                      "error": {
                        "name": "PrismaError",
                        "clientVersion": "5.22.0"
                      },
                      "status": 500
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "put": {
        "summary": "Update User",
        "deprecated": false,
        "description": "",
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": 1,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "avatar": {
                    "type": "string",
                    "description": "User Avatar",
                    "nullable": true
                  },
                  "firstName": {
                    "type": "string"
                  },
                  "lastName": {
                    "type": "string"
                  },
                  "gender": {
                    "type": "string",
                    "description": "User Gender"
                  },
                  "sex": {
                    "type": "string",
                    "enum": [
                      "male",
                      "female"
                    ],
                    "examples": [
                      "Male"
                    ]
                  },
                  "bio": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  }
                }
              },
              "example": {
                "avatar": "https://picsum.photos/seed/xw4U3l/512/512",
                "firstName": "Ethel",
                "lastName": "Bauch",
                "gender": "Bigender",
                "sex": "female",
                "bio": "engineer, teacher",
                "email": "Ethel_Bauch@hotmail.com"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/User"
                    },
                    "status": {
                      "type": "integer"
                    }
                  }
                },
                "examples": {
                  "1": {
                    "summary": "Success",
                    "value": {
                      "data": {
                        "id": 600,
                        "avatar": "https://picsum.photos/seed/aj6DaN/512/512?blur=1",
                        "firstName": "Blaze",
                        "lastName": "Conroy",
                        "gender": "Androgynous",
                        "sex": "Male",
                        "bio": "traveler",
                        "email": "Oscar_Howell27@gmail.com",
                        "createdAt": "2024-08-06T21:28:12.585Z",
                        "updatedAt": "2024-11-14T11:02:07.109Z"
                      },
                      "status": 200
                    }
                  }
                }
              }
            },
            "headers": {}
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {}
                    },
                    "status": {
                      "type": "integer"
                    }
                  }
                },
                "examples": {
                  "1": {
                    "summary": "Exception",
                    "value": {
                      "error": {
                        "issues": [
                          {
                            "code": "too_small",
                            "minimum": 2,
                            "type": "string",
                            "inclusive": true,
                            "exact": false,
                            "message": "String must contain at least 2 character(s)",
                            "path": [
                              "title"
                            ]
                          }
                        ],
                        "name": "ZodError"
                      },
                      "status": 400
                    }
                  }
                }
              }
            },
            "headers": {}
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {}
                    },
                    "status": {
                      "type": "integer"
                    }
                  }
                },
                "examples": {
                  "1": {
                    "summary": "Exception",
                    "value": {
                      "error": {
                        "name": "PrismaClientKnownRequestError",
                        "code": "P2025",
                        "clientVersion": "5.22.0",
                        "meta": {
                          "modelName": "modelName",
                          "cause": "Record not found."
                        }
                      },
                      "status": 404
                    }
                  }
                }
              }
            },
            "headers": {}
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {}
                    },
                    "status": {
                      "type": "integer"
                    }
                  }
                },
                "examples": {
                  "1": {
                    "summary": "Exception",
                    "value": {
                      "error": {
                        "name": "PrismaError",
                        "clientVersion": "5.22.0"
                      },
                      "status": 500
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Id",
            "minimum": 1
          },
          "avatar": {
            "type": "string",
            "description": "User Avatar",
            "nullable": true
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "gender": {
            "type": "string",
            "description": "User Gender"
          },
          "sex": {
            "type": "string",
            "enum": [
              "male",
              "female"
            ],
            "examples": [
              "Male"
            ]
          },
          "bio": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": false
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Develop Env"
    },
    {
      "url": "https://test.your-api-server.com",
      "description": "Testing Env"
    }
  ],
...
}